vm timing updates

+++ src/lib_gfx/vm_core.h
@@ -168,7 +168,7 @@ typedef struct __attribute__((packed, aligned(4))){ // MUST be 32 bit aligned!
     uint16_t loop_start;
 
     uint16_t frame_number;
-    uint32_t tick;
+    uint64_t tick;
     uint32_t last_elapsed_us;
     int32_t loop_delay;
